log: fix time_t usage
On 32bit platforms with 64bit time_t, gcc -Werror-format reports:

    tools/silabs-hwping/hwping.c:370:14: error: format '%ld' expects argument of type 'long int', but argument 3 has type 'time_t' {aka 'long long int'} [-Werror=format=]

    app_wsbrd/rpl/rpl_storage.c:99:48: error: format '%lu' expects argument of type 'long unsigned int', but argument 3 has type 'time_t' {aka 'long long int'} [-Werror=format=]

+++ app_wsbrd/rpl/rpl_storage.c
@@ -12,6 +12,7 @@
  * [1]: https://www.silabs.com/about-us/legal/master-software-license-agreement
  */
 #include <arpa/inet.h>
+#include <inttypes.h>
 #include <fcntl.h>
 #include <fnmatch.h>
 #include <glob.h>
@@ -96,8 +97,8 @@ void rpl_storage_store_target(const struct rpl_root *root, const struct rpl_targ
     tstamp = target->path_seq_tstamp_s + time_get_storage_offset_s();
     str_date(tstamp, time_str);
     fprintf(nvm->file, "# %s\n", time_str);
-    fprintf(nvm->file, "path_seq_timestamp = %lu\n",
-            target->path_seq_tstamp_s + time_get_storage_offset_s());
+    fprintf(nvm->file, "path_seq_timestamp = %"PRIu64"\n",
+            (uint64_t)target->path_seq_tstamp_s + time_get_storage_offset_s());
     fprintf(nvm->file, "external = %u\n", target->external);
     for (uint8_t i = 0; i < root->pcs + 1; i++) {
         if (!memzcmp(target->transits + i, sizeof(struct rpl_transit)))
